requirements.txt.txt
openai-whisper
ollama
torch
fpdf